Clarify and simplify handling of rpms and repos
Collect in one place and document the _repo and _rpms variables
in compute-vars.

Also, don't amend lists in the arch and target files, but add
specific target_* and arch_* variables.

Drop the base_rpms and boot_rpms separation, which we no longer use.

Compute image_rpms in compute-vars, instead of the mode include file.