distro: set the repo priority and make autosd.ipp.yml a symlink
Set the yum repo priority so it try to install the packages
from the AutoSD compose, if possible.
It prevents the builds to pull packages with different versions
from what the AutoSD compose have.
The change also makes the distro/autosd.ipp.yml a symlink to the latest
autosd distro definition, so it's not necessary to keep syncing.