Test more combinations in test-compose
This adds more target (rcar_s4, rpi4 and abootqemu). It also removes
the generic "use_aboot=true" define as I don't think that is going
to generally always work. We instead test it because some of the
tested targets try it.

Also, drop the custom test_qm manifest now as we don't need it anymore.