Set secure_mode_policyload selinux boolean in image mode
This adds some scaffolding so that we can set selinux booleans at
boot, and then we set secure_mode_policyload to on if
secure_mode_policyload is set, which is on by default for image mode.

This means that in image mode we cannot load a new policy after boot,
which is what we want (as policy update would only happen in a OTA update).