Support simplified manifest format
This adds support for a new, minimal *.aib.yml manifest format. Here is
the a small example that does everything default except it adds a vim
rpm:

```
name: simple

content:
  rpms:
    - vim
```

The way it works internally is that the manifest basically just
defines various mpp variables (but in a nicer layouyt) and then we
expand these in a built-in mpp file.

The code supports both *.mpp.yml files, which work as before, and
the new one, which is used if the file ends in `.aib.yml`.

The file format is defined using json-schema which makes it very clear
exactly what is allowed, and anything else will give an error. It also
allows us to auto-generate the manifest format docs using "make
generate-manifest-doc".

There are some examples in the "examples" subdir.

NOTE: We also remove the recent support for using "-" to mean
an empty mpp file that only uses mpp variables. That usecase
is now taken by the simplified format.