automotive-image-runner: Use virtio-mmio based devices for --aboot
It seems that the pci based virtio devices (disk, network) don't work
when using the device-tree based bios.

Also, virtio_mmio is currently a module, so force-load it.

Ideally we should figure out how to make the PCI work, but for now
this at least makes the disk and network work again.