Fix /var in QM partition
This does:

Copy /var from qm rootfs to /var/qm so that it is available in the
final image. Note: In image mode this just makes it available to the
ostree fillvar stage that sets up the tmpdir.d files to create the
/var structure on first deploy.

Remove /var/lib/selinux in QM partition (as QM is not managing SELinux
anyway).

In QM, move /var/lib/rpm to /usr/share/rpm and replace with a
symlink. This allows rpm -q to work in the qm parition.