Fix handling of uids, both in QM and host
There are some problems with how we use uids (and gids) in a-i-b.

First of all, any rpms that create dynamic users can get different
uids in the image depending on the order of rpm installation.  This is
problematic, because if such users e.g. own a file outside the image,
say in /var, then after an update the username may not have the same
uid.

There isn't a total solution here, but what bootc does is at least use
a fixed base /etc/passwd and /etc/group for the basic system users so
they are always the same, and then they recommend using static uid
allocation for third party users. We do the same, introducing a static
base copy of these files. And the osbuild stages to add users are always
static, so they are safe.

Secondly, the uids are currently shared between QM and rootfs, because
we're not using a user namespace (and we can't currently enable one).
This is generally fine, but it can be very confusing if the uids for
the two partitions differ. Therefore we now start the QM one with the
static base versions, and copy the final QM etc (after any qm additions)
to the rootfs version. This causes the user sets to be synchronized
and merged, which is what we need.

Additionally, the merging of the QM etc to the rootfs fixes an issue
we have with the qm /var tmpfiles.d handling which was erroring out
when trying to find the username for a uid that only existed in the
qm passwd.