On distros without linux-firmware-automotive don't install any firmwares
We used to install linux-firmware on these distros, believing that some
version was needed to satisfy dependencies. However, it turns out
that the dependencies are optional.

linux-firmware is 370 megabytes on disk and mostly contains firmware
for x86 desktop hardware. We don't need any of these on typical
automotive hardware, so its better to just skip it.

So, we default to nothing, but in cs9 we can use the
linux-firmware-automotive package from the sig, which has some
firmware that is useful for the raspberry pi.

If some specific hardware or target *do* need something from
linux-firmware, nothing prohibits installing it explicitly.