Reimplement denylist_modules and add allowlist_modules
This reimplements denylist_modules on top of the new
org.osbuild-auto.kernel.remove-modules stage. So, rather than doing
configuration changes to not load the disabled modules we remove them
from the image altogether. This means the removed modules will not be
signed, so (in addition to not being available) the kernel is not even
allowed to loads the modules (if they are accessible in some other
way).

In addition, this adds a new allowlist_modules, which removes *all*
non-listed modules (and dependencies). This optionally allows for
creation of very minimal images. For example we can avoid allowing loading
of drivers for hardware not available on the target, minimizing the risk of
potential problems from those drivers.