Support separate /var/qm partition and external /var partitions
To avoid interference via out-of-diskspace on /var from processes in
the qm partition, allow /var/qm to be on its own partition.

Also allow configuration of "external" sources for /var and
/var/qm. For these you just configure an uuid and the image will then
try to mount the partition with that uuid. This is useful in two cases,
either when the /var partition is on a completely different disk, or
when we are --export ext4 or similar that builds just a single
parition. In these cases it is up to the integrator to ensure a
partition exist with the supplied uuid.

This also removes the old use_separate_var boolean. Just set a size of
a partition to enable it.

We also add support to the simple format, by adding a "var_qm" section
similar to the "var" one in the partition dict.

Also, update size patterns to allow missing space like "20MB".