uuids: Allow override individual uuids even when using dynamic uuids
We now just have a list of defaults static uuids ones, but they are not used
if you set the corresponding variable (like rootfs_uuid)
manually. Only if you don't set it, we handle the static_uuids check.

This means it is now possible to use dynamic uuids in general but then
override a specific one to be a fixed static value.