Add integration tests
This adds support for doing full feature test of simple image format
manifests. We expect one test case per feature we want to test, which will
then build the image and look at the resulting artifact and validate that
the option was applied.

The two initial tests both use the tar export target which means we
don't have to create full images, but I expect some tests to have to build
real images too. Hopefully not to many, as that takes more time.