Make it easier for tests to create reproducible manifests
This adds a variable, reproducible_image that removes the
non-reproducible things like image_uuid and timestamp from
the generated manifests. These are great in real use to tell
the difference between real life builds, but they make the osbuild
infra to cache partial builds not work so great when tests
build a lot of images.

Also, change test-compose to use this instead of manually implementing
it.