Restructure the cs9 based distro
Rather than having the cs9 distro, which is a weird name, we split it
into two new distros. autosd9-sig, and autosd9-latest-sig.

autosd9-latest-sig is essentially what cs9 was, i.e. latest cs9,
autosd repos and the automotive sig repo.

autosd9-sig is instead the nightly autosd compose like autosd9, but
with the automotive sig repo added. And also, due to it being
available in the sig repo this uses linux-firmware-automotive.

cs9 is left as a back compat symlink (for now) to autosd9-latest-sig.

The new default distro is "autosd9-sig", but this may change in the
future.