Only install the used selinux policy
The commit 3e38ce128099b37279bcc9355646c6b0129bc773 was supposed
to move the addition of selinux-policy-$selinux_policy to base_rpms
to computed-vars.ipp, where we know the final value of $selinux_policy.

Unfortunately it forgot to remove the original reference in defaults.ipp,
which means we *also* always install targeted. Remove the copy in defaults
to fix this.